# Conflicts:
#	6_promwad2_step_by_step/H753_new/Core/Src/Call_Control/callwaiting.cpp

+++ 6_promwad2_step_by_step/H753_new/Core/Src/Call_Control/callwaiting.cpp
@@ -28,7 +28,7 @@ CallWaiting::CallWaiting()
     }
 }
 
-extern int volatile asteriskPressed;
+extern uint8_t volatile asteriskPressed;
 bool asteriskRecall = false;
 extern char msg[50];
 void CallWaiting::handleButton()
